# docs/swagger.yaml
basePath: /v1
definitions:
...
  entity.Category:
    properties:
      id:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
    type: object
  entity.CategoryWithSubCategories:
    properties:
      id:
        type: string
      name:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
      subcategories:
        items:
          $ref: '#/definitions/entity.SubCategory'
        type: array
    type: object
...
  entity.Source:
    properties:
      id:
        type: string
      site_image_url:
        type: string
      site_name:
        type: string
      site_url:
        type: string
    type: object
  entity.SubCategory:
    properties:
      category_id:
        type: string
      id:
        type: string
      name:
        type: string
      name_ru:
        type: string
      name_uz:
        type: string
    type: object
...
  models.Message:
    properties:
      message:
        type: string
    type: object
...
  v1.response:
    properties:
      error:
        example: message
        type: string
      status:
        type: boolean
    type: object
...
info:
  contact: {}
  description: 'Backend - Nodirbek No''monov     TG: https://t.me/alwaysgolang'
  title: news back-end
  version: "1.0"
paths:
...
  /categories:
    get:
      description: Retrieve all categories
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Category'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all categories
      tags:
      - Category
  /category:
    post:
      consumes:
      - application/json
      description: Create a new category with its associated subcategories in Uzbek
        and Russian
      parameters:
      - description: Category with subcategories data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.CategoryWithSubCategories'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/entity.CategoryWithSubCategories'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
      summary: Create a new category with subcategories
      tags:
      - Category
  /category/{id}:
    delete:
      description: Delete category by ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete a category
      tags:
      - Category
    get:
      consumes:
      - application/json
      description: get category by ID with Uzbek and Russian names
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: get an existing category
      tags:
      - Category
    put:
      consumes:
      - application/json
      description: Update category by ID with Uzbek and Russian names
      parameters:
      - description: Updated category data
        in: body
        name: category
        required: true
        schema:
          $ref: '#/definitions/entity.CategoryWithSubCategories'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update an existing category
      tags:
      - Category
  /category/categories:
    get:
      consumes:
      - application/json
      description: This method retrieves all categories with their subcategories,
        based on language query parameter
      operationId: getall-categories-with-subcategories
      parameters:
      - description: Language of the category (e.g. 'uz' or 'ru')
        in: query
        name: language
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.CategoryWithSubCategories'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all categories with subcategories
      tags:
      - category
  /category/source:
    get:
      consumes:
      - application/json
      description: This method gets all sources
      operationId: getall-source
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Source'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: GetAllSource
      tags:
      - source
    post:
      consumes:
      - application/json
      description: This method creates a new source
      operationId: create-source
      parameters:
      - description: Source body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/entity.Source'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Message'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      security:
      - BearerAuth: []
      summary: Create Source
      tags:
      - source
  /category/source/delete/{id}:
    delete:
      consumes:
      - application/json
      description: This method deletes given source by its id
      operationId: delete-source
      parameters:
      - description: Id of source
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.Source'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete Source
      tags:
      - source
...
  /subcategories/{id}:
    get:
      description: Retrieve all subcategories for a given category ID
      parameters:
      - description: Category ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/entity.SubCategory'
            type: array
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.response'
      summary: Get all subcategories for a category
      tags:
      - Subcategory
  /subcategory/{id}:
    delete:
      description: Delete subcategory by ID
      parameters:
      - description: Subcategory ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: Delete a subcategory
      tags:
      - Subcategory
    put:
      consumes:
      - application/json
      description: Update subcategory by ID with Uzbek and Russian names
      parameters:
      - description: Subcategory ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated subcategory data
        in: body
        name: subcategory
        required: true
        schema:
          $ref: '#/definitions/entity.SubCategory'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: Update an existing subcategory
      tags:
      - Subcategory
  /subcategory/append:
    put:
      consumes:
      - application/json
      description: Append multiple subcategories with Uzbek and Russian names
      parameters:
      - description: Array of subcategories data
        in: body
        name: subcategories
        required: true
        schema:
          items:
            $ref: '#/definitions/entity.SubCategory'
          type: array
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/v1.response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/v1.response'
      summary: Append to existing subcategories
      tags:
      - Subcategory
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"